+++ service/src/controllers/authController.ts
@@ -1,24 +1,13 @@
-import { createOctokitClientByToken, loginWithCode } from '../framework/github';
+import { loginWithCode } from '../framework/github';
 import { DEFAULT_SESSION_AGE_SECONDS } from '../consts/auth';
 
 import type { RouteHandlerMethod } from 'fastify';
 import type { FastifyValidatedRoute, LoginRequestSchema } from '../types/schemas';
-import type { UserSessionData } from '../types/auth';
 
 export const loginController: FastifyValidatedRoute<LoginRequestSchema> = async (req) => {
   const { code } = req.body;
 
-  const { token, expiresAt } = await loginWithCode(code);
-  const octokit = createOctokitClientByToken(token);
-  const { data: ghUser } = await octokit.users.getAuthenticated();
-
-  const sessionData: UserSessionData = {
-    provider: 'github',
-    name: ghUser.login,
-    auth: {
-      token,
-    },
-  };
+  const { sessionData, expiresAt } = await loginWithCode(code);
 
   req.session.options({ expires: expiresAt ?? new Date(new Date().getTime() + 1000 * DEFAULT_SESSION_AGE_SECONDS) });
   req.session.set('user', sessionData);
